Fix issue with Message.lua causing bot to crash (#290)
* Fix issue with Message.lua causing bot to crash

This fixes an issue with my friend's bot where `reaction` is nil and causes the bot to crash. The check should be before the definition of `reaction` to fix the issue.

* Apply fix for issue properly

I originally messed up and put the if statement higher when `reactions` was nil, and `reactions` was attempted to be indexed when nil, not `reaction` being nil.

* Apply requested changes

Makes the check for reactions an if statement rather than in-line with the variable.